+++ packages/create/src/cli/add/index.ts
@@ -3,6 +3,7 @@ import { addAddMockserverConfigCommand } from './mockserver-config';
 import { addAddSmartLinksConfigCommand } from './smartlinks-config';
 import { addAddCdsPluginUi5Command } from './cds-plugin-ui';
 import { addInboundNavigationConfigCommand } from './navigation-config';
+import { addCardsEditorConfigCommand } from './cards-editor';
 
 /**
  * Return 'create-fiori add *' commands. Commands include also the handler action.
@@ -19,5 +20,7 @@ export function getAddCommands(): Command {
     addAddCdsPluginUi5Command(addCommands);
     // create-fiori add inbound-navigation-config
     addInboundNavigationConfigCommand(addCommands);
+    // create-fiori add cards-editor
+    addCardsEditorConfigCommand(addCommands);
     return addCommands;
 }